server/pyproject.toml
[tool.poetry]
name = "redoodle-server"
version = "0.1.0"
description = ""
authors = []
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

fastapi = { version = "^0.115", extras = ["standard"] }
loguru = { version = "^0.7" }
Pillow = { version = "^11.0" }
timm = { version = "^1.0"}

[tool.poetry.scripts]
add-puzzles = "scripts.add_puzzles:main"

[tool.poetry.group.lint.dependencies]
ruff = "*"

[tool.poetry.group.typos.dependencies]
typos = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["redoodle_server"]

[tool.ruff.lint]
select = [
    "F", # pyflakes
    "E", # pycodestyle
    "I", # isort
    "N", # pep8-naming
    "UP", # pyupgrade
    "RUF", # ruff
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "PTH", # flake8-use-pathlib
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
]
ignore = ["E501"]
unfixable = ["F401"]

[tool.ruff.lint.isort]
force-sort-within-sections = true
split-on-trailing-comma = false
known-first-party = ["redoodle_server"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"